Refactor Distribution -> TorchDistribution class hierarchy
