+++ pyro/distributions/__init__.py
@@ -15,7 +15,7 @@
 from __future__ import absolute_import, division, print_function
 
 from pyro.distributions.delta import Delta
-from pyro.distributions.distribution import Distribution, Reshape
+from pyro.distributions.distribution import Distribution
 from pyro.distributions.rejector import Rejector
 from pyro.distributions.torch import (Bernoulli, Beta, Binomial, Categorical, Cauchy, Dirichlet, Exponential, Gamma,
                                       LogNormal, Multinomial, Normal, OneHotCategorical, Poisson,
@@ -23,5 +23,6 @@ from pyro.distributions.torch import (Bernoulli, Beta, Binomial, Categorical, Ca
 from pyro.distributions.torch.iaf import InverseAutoregressiveFlow
 from pyro.distributions.torch.multivariate_normal import MultivariateNormal
 from pyro.distributions.torch.sparse_multivariate_normal import SparseMultivariateNormal
+from pyro.distributions.torch_distribution import TorchDistribution
 
 # flake8: noqa
