+++ pyro/distributions/torch/__init__.py
@@ -5,57 +5,57 @@ import numbers
 import torch
 from torch.autograd import Variable
 
-from pyro.distributions.distribution import Distribution
+from pyro.distributions.torch_distribution import TorchDistributionMixin
 
 
-class Bernoulli(Distribution, torch.distributions.Bernoulli):
+class Bernoulli(torch.distributions.Bernoulli, TorchDistributionMixin):
     def __init__(self, ps=None, logits=None):
         super(Bernoulli, self).__init__(probs=ps, logits=logits)
 
 
-class Beta(Distribution, torch.distributions.Beta):
+class Beta(torch.distributions.Beta, TorchDistributionMixin):
     def __init__(self, alpha, beta):
         super(Beta, self).__init__(alpha, beta)
 
 
-class Binomial(Distribution, torch.distributions.Binomial):
+class Binomial(torch.distributions.Binomial, TorchDistributionMixin):
     def __init__(self, n, ps):
         super(Binomial, self).__init__(n, ps)
 
 
-class Categorical(Distribution, torch.distributions.Categorical):
+class Categorical(torch.distributions.Categorical, TorchDistributionMixin):
     def __init__(self, ps=None, logits=None):
         super(Categorical, self).__init__(probs=ps, logits=logits)
 
 
-class Cauchy(Distribution, torch.distributions.Cauchy):
+class Cauchy(torch.distributions.Cauchy, TorchDistributionMixin):
     def __init__(self, mu, gamma):
         super(Cauchy, self).__init__(mu, gamma)
 
 
-class Dirichlet(Distribution, torch.distributions.Dirichlet):
+class Dirichlet(torch.distributions.Dirichlet, TorchDistributionMixin):
     def __init__(self, alpha):
         super(Dirichlet, self).__init__(alpha)
 
 
-class Exponential(Distribution, torch.distributions.Exponential):
+class Exponential(torch.distributions.Exponential, TorchDistributionMixin):
     def __init__(self, lam):
         super(Exponential, self).__init__(lam)
 
 
-class Gamma(Distribution, torch.distributions.Gamma):
+class Gamma(torch.distributions.Gamma, TorchDistributionMixin):
     def __init__(self, alpha, beta):
         super(Gamma, self).__init__(alpha, beta)
 
 
-class LogNormal(Distribution, torch.distributions.LogNormal):
+class LogNormal(torch.distributions.LogNormal, TorchDistributionMixin):
     def __init__(self, mu, sigma):
         self.mu = mu
         self.sigma = sigma
         super(LogNormal, self).__init__(mu, sigma)
 
 
-class Multinomial(Distribution, torch.distributions.Multinomial):
+class Multinomial(torch.distributions.Multinomial, TorchDistributionMixin):
     def __init__(self, ps, n=1):
         if isinstance(n, Variable):
             n = n.data
@@ -67,25 +67,25 @@ class Multinomial(Distribution, torch.distributions.Multinomial):
         super(Multinomial, self).__init__(n, probs=ps)
 
 
-class Normal(Distribution, torch.distributions.Normal):
+class Normal(torch.distributions.Normal, TorchDistributionMixin):
     def __init__(self, mu, sigma):
         super(Normal, self).__init__(mu, sigma)
 
 
-class OneHotCategorical(Distribution, torch.distributions.OneHotCategorical):
+class OneHotCategorical(torch.distributions.OneHotCategorical, TorchDistributionMixin):
     def __init__(self, ps=None, logits=None):
         super(OneHotCategorical, self).__init__(probs=ps, logits=logits)
 
 
-class Poisson(Distribution, torch.distributions.Poisson):
+class Poisson(torch.distributions.Poisson, TorchDistributionMixin):
     def __init__(self, lam):
         super(Poisson, self).__init__(lam)
 
 
-class TransformedDistribution(Distribution, torch.distributions.TransformedDistribution):
+class TransformedDistribution(torch.distributions.TransformedDistribution, TorchDistributionMixin):
     pass
 
 
-class Uniform(Distribution, torch.distributions.Uniform):
+class Uniform(torch.distributions.Uniform, TorchDistributionMixin):
     def __init__(self, a, b):
         super(Uniform, self).__init__(a, b)
